tt_scraping_test/requirements.txt
apify-client==1.6.0
pandas==2.2.2
requests==2.32.3
python-decouple==3.8